API/requirements.txt
Flask==2.2.3
pycryptodome==3.16.0